chatbot/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
mcp[cli]